add summarize_train
