+++ alf/algorithms/config.py
@@ -186,3 +186,49 @@ class TrainerConfig(object):
             num_envs=num_envs)
         for k, v in parameters.items():
             self.__setattr__(k, v)
+
+
+@gin.configurable
+class SupervisedTrainerConfig(object):
+    """Configuration for supervised training."""
+
+    def __init__(self,
+                 root_dir,
+                 algorithm_ctor=None,
+                 random_seed=None,
+                 epochs=2e+5,
+                 eval_accuracy=True,
+                 eval_uncertainty=False,
+                 summary_interval=50,
+                 summaries_flush_secs=1,
+                 summary_max_queue=100,
+                 debug_summaries=False,
+                 summarize_grads_and_vars=False):
+        """
+        Args:
+            root_dir (str): directory for saving summary and checkpoints
+            algorithm_ctor (Callable): callable that create an
+                ``OffPolicyAlgorithm`` or ``OnPolicyAlgorithm`` instance
+            random_seed (None|int): random seed, a random seed is used if None
+            epochs (int): number of training epoches
+            eval_accuracy (bool): whether to evluate accuracy after training
+            eval_uncertainty (bool): whether to evluate uncertainty after training
+            summary_interval (int): write summary every so many training steps
+            summaries_flush_secs (int): flush summary to disk every so many seconds
+            summary_max_queue (int): flush to disk every so mary summaries
+            summarize_grads_and_vars (bool): If True, gradient and network variable
+                summaries will be written during training.
+        """
+        parameters = dict(
+            root_dir=root_dir,
+            algorithm_ctor=algorithm_ctor,
+            random_seed=random_seed,
+            epochs=epochs,
+            evaluate=eval_accuracy,
+            summary_interval=summary_interval,
+            summaries_flush_secs=summaries_flush_secs,
+            summary_max_queue=summary_max_queue,
+            debug_summaries=debug_summaries,
+            summarize_grads_and_vars=summarize_grads_and_vars)
+        for k, v in parameters.items():
+            self.__setattr__(k, v)
